pyproject.toml
[project]
name = "shunyaku"
version = "0.1.0"
description = "Clipboard and OCR popup translator from English to Japanese."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Operating System :: MacOS",
  "License :: OSI Approved :: MIT License",
]
dependencies = [
  "huggingface-hub>=0.33.0",
  "llama-cpp-python>=0.3.9",
  "pynput>=1.7.7",
  "PySide6>=6.9.0",
  "pyobjc-framework-Quartz>=12.1",
  "pyobjc-framework-Vision>=12.1",
]

[project.scripts]
shunyaku = "shunyaku.app:main"

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
